Added enums for IES format
Added enums for both photometry type, which is automatically parsed from the file, and luminous opening type. The luminous opening is determined from the zero-valued, negative and positive width, length and height fields. I have added tests to ensure this is correct, both at the source and at the interface level in the IES object.

In this commit, I have also run `cargo fmt` across the codebase.