Beginning of PhotometricWeb implementation.
I have added basic implementation of intensity for the `PhotometricWeb` structure. Unit tests have been added to support this code. I have added some helper functions for the conversion of IES to a `PhotometricWeb`, but the implementation is far from complete.

I have however tried to document the process and add other documentation to the crate to allow better understanding of structure and content.