IES (Type C) file now convert to PhotometricWeb.
First implementation of conversion to `PhotometricWeb` objects from an IES file is now in place. I have made an effort to make the resolution of symmetries automatic and as neat as possible. To this end, I have added utility functions for mirroring quadrants and hemispheres. This has also neatened the EULUMDAT conversion too. I have also made sure that the `PhotometricWeb` object can be cloned.

Accompanying this commit is an example file, copied from Annex C of the IES 2002 spec, as well as tests to perform basic verification of the new code that has been added.

I have also run `cargo fmt` across the code base for recent changes.